requirements.txt
Flask==2.2.2
requests==2.31.0
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0
kafka-python==2.0.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
python-dotenv==1.0.0
